refactor: move inode models and rename handler
