+++ backend/server/src/handlers/files.rs
@@ -1,5 +1,8 @@
 use crate::error::AppError;
-use crate::models::{FileInfo, NodeType, TreeNode};
+use crate::models::{
+    NodeType, TreeNode,
+    inode::{Inode, node2list},
+};
 use crate::state::AppState;
 use axum::{
     Json,
@@ -8,37 +11,6 @@ use axum::{
     response::{IntoResponse, Response},
 };
 use percent_encoding::percent_decode_str;
-use serde::Serialize;
-
-#[derive(Serialize, Clone)]
-#[serde(tag = "type")]
-pub enum Node {
-    #[serde(rename = "file")]
-    File { name: String, info: FileInfo },
-    #[serde(rename = "folder")]
-    Folder { name: String },
-}
-
-fn node2list(node: &TreeNode) -> Vec<Node> {
-    node.iter()
-        .map(|(name, value)| match value {
-            NodeType::File(info) => Node::File {
-                name: name.clone(),
-                info: info.clone(),
-            },
-            NodeType::Node(_) => Node::Folder { name: name.clone() },
-        })
-        .collect()
-}
-
-#[derive(Serialize)]
-#[serde(tag = "type")]
-enum Inode {
-    #[serde(rename = "folder")]
-    Folder { data: Vec<Node> },
-    #[serde(rename = "file")]
-    File { name: String, info: FileInfo },
-}
 
 // Synchronous helpers used by the router closures
 pub async fn get_node(
