+++ backend/server/src/handlers/mod.rs
@@ -1,9 +1,9 @@
-pub mod inode;
+pub mod files;
 pub mod intro;
 pub mod search;
 pub mod wiki;
 
-pub use inode::{get_node, get_node_root};
+pub use files::{get_node, get_node_root};
 pub use intro::{find_name, intro};
 pub use search::{search, search_combined};
 pub use wiki::wiki_search_picture;
